requirements.txt
questionary
Kingfish
memory-profiler
psutil
speedtest-cli